+++ tests/adapters/test_task_state_common.py
@@ -1,9 +1,13 @@
 """
-Contract tests for TaskStateProtocol + TaskSubmitProtocol implementations.
+Contract tests for TaskStateProtocol implementations.
 
 Uses only the saga-path public API (the same methods StateManager calls when
 ``force_saga=True``).  No ``data_store.*`` access; no sorted-set assertions.
 Runs against all three backends (raw, json, sql) with zero xfails.
+
+The ``task_adapter`` fixture yields a ``(state, submit)`` pair.  ``submit`` is
+used only for test setup (seeding tasks via ``submit_task``); every assertion
+targets a ``TaskStateProtocol`` method on ``state``.
 """
 
 import datetime as dt
@@ -37,7 +41,7 @@ def make_task(
     return task
 
 
-# ── basic ─────────────────────────────────────────────────────────────────────
+# ── get_all_tasks ─────────────────────────────────────────────────────────────
 
 
 @pytest.mark.asyncio
@@ -53,13 +57,9 @@ async def test_get_all_tasks_returns_submitted_tasks(task_adapter):
     """A submitted task is returned by get_all_tasks for its queue."""
     state, submit = task_adapter
     await submit.submit_task(make_task())
-
     results = await state.get_all_tasks(TaskPagination(queue="default"))
     assert len(results) == 1
     assert results[0].id == ULID1
-
-
-# ── filters ───────────────────────────────────────────────────────────────────
 
 
 @pytest.mark.asyncio
@@ -68,7 +68,6 @@ async def test_get_all_tasks_filters_by_task_name(task_adapter):
     state, submit = task_adapter
     await submit.submit_task(make_task(ULID1, name="task_a"))
     await submit.submit_task(make_task(ULID2, name="task_b"))
-
     results = await state.get_all_tasks(TaskPagination(queue="default", task_name="task_a"))
     assert len(results) == 1
     assert results[0].name == "task_a"
@@ -80,7 +79,6 @@ async def test_get_all_tasks_filters_by_version(task_adapter):
     state, submit = task_adapter
     await submit.submit_task(make_task(ULID1, version=1))
     await submit.submit_task(make_task(ULID2, version=2))
-
     results = await state.get_all_tasks(TaskPagination(queue="default", task_version=1))
     assert len(results) == 1
     assert results[0].version == 1
@@ -93,7 +91,6 @@ async def test_get_all_tasks_filters_by_status(task_adapter):
     await submit.submit_task(make_task(ULID1, status=TaskStatus.SUBMITTED))
     await submit.submit_task(make_task(ULID2, status=TaskStatus.SUBMITTED))
     await state.save_task(make_task(ULID2, status=TaskStatus.COMPLETED))
-
     results = await state.get_all_tasks(TaskPagination(queue="default", status=TaskStatus.SUBMITTED))
     assert len(results) == 1
     assert results[0].id == ULID1
@@ -105,7 +102,6 @@ async def test_get_all_tasks_different_queues_are_isolated(task_adapter):
     state, submit = task_adapter
     await submit.submit_task(make_task(ULID1, queue="queue_a"))
     await submit.submit_task(make_task(ULID2, queue="queue_b"))
-
     results = await state.get_all_tasks(TaskPagination(queue="queue_a"))
     assert len(results) == 1
     assert results[0].id == ULID1
@@ -116,12 +112,8 @@ async def test_get_all_tasks_filter_no_match_returns_empty(task_adapter):
     """A filter that matches no tasks returns an empty list."""
     state, submit = task_adapter
     await submit.submit_task(make_task(ULID1, name="task_a"))
-
     results = await state.get_all_tasks(TaskPagination(queue="default", task_name="task_nonexistent"))
     assert results == []
-
-
-# ── pagination ────────────────────────────────────────────────────────────────
 
 
 @pytest.mark.asyncio
@@ -130,7 +122,6 @@ async def test_get_all_tasks_respects_limit(task_adapter):
     state, submit = task_adapter
     for i, uid in enumerate((ULID1, ULID2, ULID3)):
         await submit.submit_task(make_task(uid, submitted_at=FROZEN_TIME + dt.timedelta(seconds=i)))
-
     results = await state.get_all_tasks(TaskPagination(queue="default", limit=2))
     assert len(results) == 2
 
@@ -141,14 +132,12 @@ async def test_get_all_tasks_offset_skips_first_n(task_adapter):
     state, submit = task_adapter
     for i, uid in enumerate((ULID1, ULID2, ULID3)):
         await submit.submit_task(make_task(uid, submitted_at=FROZEN_TIME + dt.timedelta(seconds=i)))
-
     page1 = await state.get_all_tasks(
         TaskPagination(queue="default", limit=2, offset=0, order_by=PaginationOrder.SUBMITTED_AT)
     )
     page2 = await state.get_all_tasks(
         TaskPagination(queue="default", limit=2, offset=2, order_by=PaginationOrder.SUBMITTED_AT)
     )
-
     assert len(page1) == 2
     assert len(page2) == 1
     assert {t.id for t in page1}.isdisjoint({t.id for t in page2})
@@ -159,78 +148,54 @@ async def test_get_all_tasks_offset_returns_empty_beyond_end(task_adapter):
     """An offset past the total count returns an empty list."""
     state, submit = task_adapter
     await submit.submit_task(make_task(ULID1))
-
     results = await state.get_all_tasks(TaskPagination(queue="default", limit=10, offset=5))
     assert results == []
 
 
 @pytest.mark.asyncio
 async def test_get_all_tasks_order_by_submitted_at(task_adapter):
-    """
-    Results are returned in ascending submitted_at order within each page.
-
-    Tasks are submitted newest-first to prove the sort is applied rather than
-    relying on insertion order.  With 5 tasks and limit=3, page 1 returns the
-    3 oldest in order and page 2 returns the 2 newest in order.
-    """
+    """Results are returned in ascending submitted_at order within each page."""
     state, submit = task_adapter
     tasks = [
         make_task(uid, submitted_at=FROZEN_TIME + dt.timedelta(seconds=i))
         for i, uid in enumerate((ULID1, ULID2, ULID3, ULID4, ULID5))
     ]
-    for t in reversed(tasks):  # submit newest-first to prove sort is applied
+    for t in reversed(tasks):
         await submit.submit_task(t)
-
     page1 = await state.get_all_tasks(
         TaskPagination(queue="default", limit=3, offset=0, order_by=PaginationOrder.SUBMITTED_AT)
     )
     page2 = await state.get_all_tasks(
         TaskPagination(queue="default", limit=3, offset=3, order_by=PaginationOrder.SUBMITTED_AT)
     )
-
     assert [t.id for t in page1] == [ULID1, ULID2, ULID3]
     assert [t.id for t in page2] == [ULID4, ULID5]
 
 
 @pytest.mark.asyncio
 async def test_get_all_tasks_order_by_task_id(task_adapter):
-    """
-    order_by=TASK_ID returns tasks sorted by ULID.
-
-    Tasks are submitted with ULID order matching submitted_at order so the test
-    passes for all adapter implementations (RedisTaskState sorts by
-    submitted_at regardless of order_by, which coincides here).
-    Covers the non-SUBMITTED_AT code path in RedisJSONTaskState.get_all_tasks.
-    """
+    """order_by=TASK_ID returns tasks sorted by ULID."""
     state, submit = task_adapter
     t1 = make_task(ULID1, submitted_at=FROZEN_TIME)
     t2 = make_task(ULID2, submitted_at=FROZEN_TIME + dt.timedelta(seconds=1))
     t3 = make_task(ULID3, submitted_at=FROZEN_TIME + dt.timedelta(seconds=2))
     for t in (t1, t2, t3):
         await submit.submit_task(t)
-
-    results = await state.get_all_tasks(
-        TaskPagination(queue="default", order_by=PaginationOrder.TASK_ID)
-    )
+    results = await state.get_all_tasks(TaskPagination(queue="default", order_by=PaginationOrder.TASK_ID))
     assert [t.id for t in results] == [ULID1, ULID2, ULID3]
-
-
-# ── combined filters ──────────────────────────────────────────────────────────
 
 
 @pytest.mark.asyncio
 async def test_get_all_tasks_combined_name_and_version(task_adapter):
     """Filtering by both task_name and task_version returns only exact matches."""
     state, submit = task_adapter
-    t1 = make_task(ULID1, name="task_a", version=1)
-    t2 = make_task(ULID2, name="task_a", version=2)
-    t3 = make_task(ULID3, name="task_b", version=1)
-    for t in (t1, t2, t3):
+    for t in (
+        make_task(ULID1, name="task_a", version=1),
+        make_task(ULID2, name="task_a", version=2),
+        make_task(ULID3, name="task_b", version=1),
+    ):
         await submit.submit_task(t)
-
-    results = await state.get_all_tasks(
-        TaskPagination(queue="default", task_name="task_a", task_version=1)
-    )
+    results = await state.get_all_tasks(TaskPagination(queue="default", task_name="task_a", task_version=1))
     assert len(results) == 1
     assert results[0].id == ULID1
 
@@ -239,12 +204,9 @@ async def test_get_all_tasks_combined_name_and_version(task_adapter):
 async def test_get_all_tasks_combined_name_and_status(task_adapter):
     """Filtering by both task_name and status returns only tasks matching both."""
     state, submit = task_adapter
-    t1 = make_task(ULID1, name="task_a", status=TaskStatus.SUBMITTED)
-    t2 = make_task(ULID2, name="task_a", status=TaskStatus.SUBMITTED)
-    await submit.submit_task(t1)
-    await submit.submit_task(t2)
+    await submit.submit_task(make_task(ULID1, name="task_a", status=TaskStatus.SUBMITTED))
+    await submit.submit_task(make_task(ULID2, name="task_a", status=TaskStatus.SUBMITTED))
     await state.save_task(make_task(ULID2, name="task_a", status=TaskStatus.COMPLETED))
-
     results = await state.get_all_tasks(
         TaskPagination(queue="default", task_name="task_a", status=TaskStatus.SUBMITTED)
     )
@@ -256,12 +218,9 @@ async def test_get_all_tasks_combined_name_and_status(task_adapter):
 async def test_get_all_tasks_combined_version_and_status(task_adapter):
     """Filtering by version and status without task_name works correctly."""
     state, submit = task_adapter
-    t1 = make_task(ULID1, version=1, status=TaskStatus.SUBMITTED)
-    t2 = make_task(ULID2, version=1, status=TaskStatus.SUBMITTED)
-    await submit.submit_task(t1)
-    await submit.submit_task(t2)
+    await submit.submit_task(make_task(ULID1, version=1, status=TaskStatus.SUBMITTED))
+    await submit.submit_task(make_task(ULID2, version=1, status=TaskStatus.SUBMITTED))
     await state.save_task(make_task(ULID2, version=1, status=TaskStatus.FAILED))
-
     results = await state.get_all_tasks(
         TaskPagination(queue="default", task_version=1, status=TaskStatus.SUBMITTED)
     )
@@ -273,35 +232,23 @@ async def test_get_all_tasks_combined_version_and_status(task_adapter):
 async def test_get_all_tasks_all_filters_combined(task_adapter):
     """All three filters together (name, version, status) are applied conjunctively."""
     state, submit = task_adapter
-    t1 = make_task(ULID1, name="task_a", version=1, status=TaskStatus.SUBMITTED)
-    t2 = make_task(ULID2, name="task_a", version=2, status=TaskStatus.SUBMITTED)
-    t3 = make_task(ULID3, name="task_b", version=1, status=TaskStatus.SUBMITTED)
-    for t in (t1, t2, t3):
+    for t in (
+        make_task(ULID1, name="task_a", version=1, status=TaskStatus.SUBMITTED),
+        make_task(ULID2, name="task_a", version=2, status=TaskStatus.SUBMITTED),
+        make_task(ULID3, name="task_b", version=1, status=TaskStatus.SUBMITTED),
+    ):
         await submit.submit_task(t)
     await state.save_task(make_task(ULID1, name="task_a", version=1, status=TaskStatus.COMPLETED))
-
     results = await state.get_all_tasks(
         TaskPagination(queue="default", task_name="task_a", task_version=1, status=TaskStatus.SUBMITTED)
     )
     assert results == []
 
 
-# ── filter + pagination ───────────────────────────────────────────────────────
-
-
 @pytest.mark.asyncio
 async def test_get_all_tasks_filter_pagination_page_size_is_predictable(task_adapter):
     """
-    Page size is predictable when a filter is active (desired contract).
-
-    When offset counts matched documents (not raw queue positions), each page returns
-    exactly ``limit`` results (or fewer only when fewer matching tasks remain), and
-    consecutive pages together cover all matching tasks without gaps or duplicates.
-
-    Setup: 4 tasks submitted earliest-first — B1, B2, A1, A2.  Only A* match the filter.
-    With limit=1 and filter task_name=task_a:
-      page 1 (offset=0) → exactly 1 result (A1)
-      page 2 (offset=1) → exactly 1 result (A2, not A1 again)
+    Page size is predictable when a filter is active.
 
     Known limitation: RedisTaskState applies offset to raw queue positions before
     Python-side filtering, so offset-based pagination over a filtered result set is
@@ -313,83 +260,37 @@ async def test_get_all_tasks_filter_pagination_page_size_is_predictable(task_ada
             "RedisTaskState applies offset before Python-side filtering; "
             "page 2 backs up into queue-space and returns the same task as page 1"
         )
-
-    b1 = make_task(ULID1, name="task_b", submitted_at=FROZEN_TIME)
-    b2 = make_task(ULID2, name="task_b", submitted_at=FROZEN_TIME + dt.timedelta(seconds=1))
-    a1 = make_task(ULID3, name="task_a", submitted_at=FROZEN_TIME + dt.timedelta(seconds=2))
-    a2 = make_task(ULID4, name="task_a", submitted_at=FROZEN_TIME + dt.timedelta(seconds=3))
-    for t in (b1, b2, a1, a2):
+    for t in (
+        make_task(ULID1, name="task_b", submitted_at=FROZEN_TIME),
+        make_task(ULID2, name="task_b", submitted_at=FROZEN_TIME + dt.timedelta(seconds=1)),
+        make_task(ULID3, name="task_a", submitted_at=FROZEN_TIME + dt.timedelta(seconds=2)),
+        make_task(ULID4, name="task_a", submitted_at=FROZEN_TIME + dt.timedelta(seconds=3)),
+    ):
         await submit.submit_task(t)
-
     page1 = await state.get_all_tasks(
         TaskPagination(
-            queue="default",
-            task_name="task_a",
-            limit=1,
-            offset=0,
-            order_by=PaginationOrder.SUBMITTED_AT,
+            queue="default", task_name="task_a", limit=1, offset=0, order_by=PaginationOrder.SUBMITTED_AT
         )
     )
     page2 = await state.get_all_tasks(
         TaskPagination(
-            queue="default",
-            task_name="task_a",
-            limit=1,
-            offset=1,
-            order_by=PaginationOrder.SUBMITTED_AT,
+            queue="default", task_name="task_a", limit=1, offset=1, order_by=PaginationOrder.SUBMITTED_AT
         )
     )
-
     assert [t.id for t in page1] == [ULID3]
     assert [t.id for t in page2] == [ULID4]
 
 
-# ── submit_task / get_task / task_exists ──────────────────────────────────────
-
-
-@pytest.mark.asyncio
-async def test_submit_task(task_adapter):
-    """submit_task enqueues the task and persists its data."""
-    state, submit = task_adapter
-    task = Task(id=ULID1, name="Test Task", status=TaskStatus.UNSUBMITTED, queue="default")
-    task.set_status(TaskStatus.SUBMITTED)
-    await submit.submit_task(task)
-
-    saved = await state.get_task(ULID1)
-    assert saved is not None
-    assert saved.name == "Test Task"
-    assert saved.status == TaskStatus.SUBMITTED
-    assert saved.submitted_at == task.submitted_at
-    assert await state.task_exists(ULID1)
-
-
-@pytest.mark.asyncio
-async def test_submit_task_twice_updates_only(task_adapter):
-    """Submitting the same task ID twice updates the data without duplicating the queue entry."""
-    state, submit = task_adapter
-    task = Task(id=ULID1, name="Initial Task", status="unsubmitted")
-    task.set_status(TaskStatus.SUBMITTED)
-    await submit.submit_task(task)
-
-    updated_task = Task(id=ULID1, name="Updated Task", status="completed", submitted_at=task.submitted_at)
-    await submit.submit_task(updated_task)
-
-    saved = await state.get_task(ULID1)
-    assert saved is not None
-    assert saved.name == "Updated Task"
-    assert saved.status == TaskStatus.COMPLETED
-    assert saved.submitted_at == task.submitted_at
-
-    results = await state.get_all_tasks(TaskPagination(queue=task.queue))
-    assert len(results) == 1
+# ── get_task / task_exists ────────────────────────────────────────────────────
 
 
 @pytest.mark.asyncio
 async def test_get_task(task_adapter):
     """get_task returns the saved task with all fields intact."""
     state, submit = task_adapter
-    task_to_save = Task(id=ULID1, name="Test Task", status=TaskStatus.STARTED, submitted_at=FROZEN_TIME)
-    await state.save_task(task_to_save)
+    await state.save_task(
+        Task(id=ULID1, name="Test Task", status=TaskStatus.STARTED, submitted_at=FROZEN_TIME)
+    )
     task = await state.get_task(ULID1)
     assert task is not None
     assert task.id == ULID1
@@ -421,15 +322,14 @@ async def test_task_exists(task_adapter):
 
 @pytest.mark.asyncio
 async def test_update_task_heartbeat_sets_timestamp(task_adapter):
-    """update_task_heartbeat is reflected in the heartbeat_at field when the task is retrieved."""
+    """update_task_heartbeat is reflected in heartbeat_at when the task is retrieved."""
     state, submit = task_adapter
     task = make_task(status=TaskStatus.STARTED)
     await state.save_task(task)
     task.heartbeat_at = FROZEN_TIME
     await state.update_task_heartbeat(task)
-
-    updated_task = await state.get_task(task.id)
-    assert updated_task.heartbeat_at == task.heartbeat_at
+    updated = await state.get_task(task.id)
+    assert updated.heartbeat_at == FROZEN_TIME
 
 
 @pytest.mark.asyncio
@@ -440,9 +340,7 @@ async def test_update_task_heartbeat_makes_task_active(task_adapter):
     await state.save_task(task)
     task.heartbeat_at = FROZEN_TIME
     await state.update_task_heartbeat(task)
-
-    active = await state.get_active_tasks({task.queue})
-    assert any(t.id == task.id for t in active)
+    assert any(t.id == task.id for t in await state.get_active_tasks({task.queue}))
 
 
 @pytest.mark.asyncio
@@ -453,11 +351,9 @@ async def test_update_task_heartbeat_updates_existing_entry(task_adapter):
     await state.save_task(task)
     task.heartbeat_at = FROZEN_TIME
     await state.update_task_heartbeat(task)
-
     second_time = FROZEN_TIME + dt.timedelta(seconds=10)
     task.heartbeat_at = second_time
     await state.update_task_heartbeat(task)
-
     updated = await state.get_task(task.id)
     assert updated is not None
     assert updated.heartbeat_at == second_time
@@ -475,7 +371,6 @@ async def test_remove_task_heartbeat_removes_entry(task_adapter):
     task.heartbeat_at = FROZEN_TIME
     await state.update_task_heartbeat(task)
     assert any(t.id == task.id for t in await state.get_active_tasks({task.queue}))
-
     await state.remove_task_heartbeat(task)
     assert not any(t.id == task.id for t in await state.get_active_tasks({task.queue}))
 
@@ -484,8 +379,7 @@ async def test_remove_task_heartbeat_removes_entry(task_adapter):
 async def test_remove_task_heartbeat_noop_when_absent(task_adapter):
     """remove_task_heartbeat is a no-op when the task has no heartbeat entry."""
     state, submit = task_adapter
-    task = make_task()
-    await state.remove_task_heartbeat(task)  # should not raise
+    await state.remove_task_heartbeat(make_task())  # should not raise
 
 
 # ── get_active_tasks ──────────────────────────────────────────────────────────
@@ -495,8 +389,7 @@ async def test_remove_task_heartbeat_noop_when_absent(task_adapter):
 async def test_get_active_tasks_returns_empty_for_no_heartbeats(task_adapter):
     """get_active_tasks returns [] when no tasks have heartbeat entries."""
     state, submit = task_adapter
-    result = await state.get_active_tasks({"default"})
-    assert result == []
+    assert await state.get_active_tasks({"default"}) == []
 
 
 @pytest.mark.asyncio
@@ -507,7 +400,6 @@ async def test_get_active_tasks_returns_tasks_with_heartbeats(task_adapter):
     await state.save_task(task)
     task.heartbeat_at = FROZEN_TIME
     await state.update_task_heartbeat(task)
-
     result = await state.get_active_tasks({"default"})
     assert len(result) == 1
     assert result[0].id == ULID1
@@ -521,20 +413,19 @@ async def test_get_stale_tasks_returns_stale_tasks(task_adapter):
     """Tasks whose heartbeat is older than stale_time are returned."""
     state, submit = task_adapter
     now = dt.datetime.now(dt.UTC)
-    stale_task = Task(
+    stale = Task(
         id=ULID1,
-        name="stale_task",
+        name="stale",
         queue="default",
         status=TaskStatus.STARTED,
         started_at=now,
         heartbeat_at=now - dt.timedelta(minutes=10),
     )
-    await state.save_task(stale_task)
-    await state.update_task_heartbeat(stale_task)
-
-    stale_tasks = [task async for task in state.get_stale_tasks({"default"}, dt.timedelta(minutes=5))]
-    assert len(stale_tasks) == 1
-    assert stale_tasks[0].id == stale_task.id
+    await state.save_task(stale)
+    await state.update_task_heartbeat(stale)
+    result = [t async for t in state.get_stale_tasks({"default"}, dt.timedelta(minutes=5))]
+    assert len(result) == 1
+    assert result[0].id == ULID1
 
 
 @pytest.mark.asyncio
@@ -542,19 +433,18 @@ async def test_get_stale_tasks_excludes_recent_tasks(task_adapter):
     """Tasks whose heartbeat is within stale_time are not returned."""
     state, submit = task_adapter
     now = dt.datetime.now(dt.UTC)
-    recent_task = Task(
+    recent = Task(
         id=ULID1,
-        name="recent_task",
+        name="recent",
         queue="default",
         status=TaskStatus.STARTED,
         started_at=now,
         heartbeat_at=now - dt.timedelta(minutes=1),
     )
-    await state.save_task(recent_task)
-    await state.update_task_heartbeat(recent_task)
-
-    stale_tasks = [task async for task in state.get_stale_tasks({"default"}, dt.timedelta(minutes=5))]
-    assert len(stale_tasks) == 0
+    await state.save_task(recent)
+    await state.update_task_heartbeat(recent)
+    result = [t async for t in state.get_stale_tasks({"default"}, dt.timedelta(minutes=5))]
+    assert len(result) == 0
 
 
 @pytest.mark.asyncio
@@ -562,32 +452,19 @@ async def test_get_stale_tasks_handles_multiple_queues(task_adapter):
     """Stale tasks from multiple queues are all returned."""
     state, submit = task_adapter
     now = dt.datetime.now(dt.UTC)
+    hb = now - dt.timedelta(minutes=10)
     t1 = Task(
-        id=ULID1,
-        name="task_1",
-        queue="default",
-        status=TaskStatus.STARTED,
-        started_at=now,
-        heartbeat_at=now - dt.timedelta(minutes=10),
+        id=ULID1, name="t1", queue="default", status=TaskStatus.STARTED, started_at=now, heartbeat_at=hb
     )
     t2 = Task(
-        id=ULID2,
-        name="task_2",
-        queue="high_priority",
-        status=TaskStatus.STARTED,
-        started_at=now,
-        heartbeat_at=now - dt.timedelta(minutes=10),
+        id=ULID2, name="t2", queue="high_priority", status=TaskStatus.STARTED, started_at=now, heartbeat_at=hb
     )
     for t in (t1, t2):
         await state.save_task(t)
         await state.update_task_heartbeat(t)
-
-    stale_tasks = [
-        task
-        async for task in state.get_stale_tasks({"default", "high_priority"}, dt.timedelta(minutes=5))
-    ]
-    assert len(stale_tasks) == 2
-    assert {task.id for task in stale_tasks} == {t1.id, t2.id}
+    result = [t async for t in state.get_stale_tasks({"default", "high_priority"}, dt.timedelta(minutes=5))]
+    assert len(result) == 2
+    assert {t.id for t in result} == {ULID1, ULID2}
 
 
 # ── clean_terminal_tasks ──────────────────────────────────────────────────────
@@ -597,14 +474,15 @@ async def test_get_stale_tasks_handles_multiple_queues(task_adapter):
 async def test_clean_terminal_tasks_deletes_old_completed_task(task_adapter):
     """A completed task blob older than max_age is deleted."""
     state, submit = task_adapter
-    task = Task(
-        id=ULID1,
-        name="test_task",
-        queue="default",
-        status=TaskStatus.COMPLETED,
-        completed_at=FROZEN_TIME - dt.timedelta(days=8),
+    await state.save_task(
+        Task(
+            id=ULID1,
+            name="t",
+            queue="default",
+            status=TaskStatus.COMPLETED,
+            completed_at=FROZEN_TIME - dt.timedelta(days=8),
+        )
     )
-    await state.save_task(task)
     await state.clean_terminal_tasks(FROZEN_TIME, dt.timedelta(days=7))
     assert not await state.task_exists(ULID1)
 
@@ -613,14 +491,15 @@ async def test_clean_terminal_tasks_deletes_old_completed_task(task_adapter):
 async def test_clean_terminal_tasks_skips_active_task(task_adapter):
     """Tasks in active statuses are never deleted regardless of age."""
     state, submit = task_adapter
-    task = Task(
-        id=ULID1,
-        name="test_task",
-        queue="default",
-        status=TaskStatus.STARTED,
-        started_at=FROZEN_TIME - dt.timedelta(days=100),
+    await state.save_task(
+        Task(
+            id=ULID1,
+            name="t",
+            queue="default",
+            status=TaskStatus.STARTED,
+            started_at=FROZEN_TIME - dt.timedelta(days=100),
+        )
     )
-    await state.save_task(task)
     await state.clean_terminal_tasks(FROZEN_TIME, dt.timedelta(seconds=0))
     assert await state.task_exists(ULID1)
 
@@ -629,14 +508,15 @@ async def test_clean_terminal_tasks_skips_active_task(task_adapter):
 async def test_clean_terminal_tasks_skips_task_within_age(task_adapter):
     """A completed task whose completed_at is within max_age is not deleted."""
     state, submit = task_adapter
-    task = Task(
-        id=ULID1,
-        name="test_task",
-        queue="default",
-        status=TaskStatus.COMPLETED,
-        completed_at=FROZEN_TIME - dt.timedelta(hours=1),
+    await state.save_task(
+        Task(
+            id=ULID1,
+            name="t",
+            queue="default",
+            status=TaskStatus.COMPLETED,
+            completed_at=FROZEN_TIME - dt.timedelta(hours=1),
+        )
     )
-    await state.save_task(task)
     await state.clean_terminal_tasks(FROZEN_TIME, dt.timedelta(days=7))
     assert await state.task_exists(ULID1)
 
@@ -645,8 +525,9 @@ async def test_clean_terminal_tasks_skips_task_within_age(task_adapter):
 async def test_clean_terminal_tasks_skips_task_without_completed_at(task_adapter):
     """A terminal task with no completed_at is not deleted."""
     state, submit = task_adapter
-    task = Task(id=ULID1, name="test_task", queue="default", status=TaskStatus.FAILED, completed_at=None)
-    await state.save_task(task)
+    await state.save_task(
+        Task(id=ULID1, name="t", queue="default", status=TaskStatus.FAILED, completed_at=None)
+    )
     await state.clean_terminal_tasks(FROZEN_TIME, dt.timedelta(seconds=0))
     assert await state.task_exists(ULID1)
 
@@ -665,14 +546,15 @@ async def test_clean_terminal_tasks_skips_task_without_completed_at(task_adapter
 async def test_clean_terminal_tasks_cleans_all_terminal_statuses(task_adapter, status):
     """All five terminal statuses are eligible for cleanup when old enough."""
     state, submit = task_adapter
-    task = Task(
-        id=ULID1,
-        name="test_task",
-        queue="default",
-        status=status,
-        completed_at=FROZEN_TIME - dt.timedelta(days=8),
+    await state.save_task(
+        Task(
+            id=ULID1,
+            name="t",
+            queue="default",
+            status=status,
+            completed_at=FROZEN_TIME - dt.timedelta(days=8),
+        )
     )
-    await state.save_task(task)
     await state.clean_terminal_tasks(FROZEN_TIME, dt.timedelta(days=7))
     assert not await state.task_exists(ULID1)
 
@@ -681,41 +563,41 @@ async def test_clean_terminal_tasks_cleans_all_terminal_statuses(task_adapter, s
 async def test_clean_terminal_tasks_leaves_other_tasks_untouched(task_adapter):
     """Only old terminal tasks are deleted; recent tasks remain."""
     state, submit = task_adapter
-    old_task = Task(
-        id=ULID1,
-        name="test_task",
-        queue="default",
-        status=TaskStatus.COMPLETED,
-        completed_at=FROZEN_TIME - dt.timedelta(days=8),
+    await state.save_task(
+        Task(
+            id=ULID1,
+            name="t",
+            queue="default",
+            status=TaskStatus.COMPLETED,
+            completed_at=FROZEN_TIME - dt.timedelta(days=8),
+        )
     )
-    recent_task = Task(
-        id=ULID2,
-        name="test_task",
-        queue="default",
-        status=TaskStatus.COMPLETED,
-        completed_at=FROZEN_TIME - dt.timedelta(hours=1),
+    await state.save_task(
+        Task(
+            id=ULID2,
+            name="t",
+            queue="default",
+            status=TaskStatus.COMPLETED,
+            completed_at=FROZEN_TIME - dt.timedelta(hours=1),
+        )
     )
-    await state.save_task(old_task)
-    await state.save_task(recent_task)
     await state.clean_terminal_tasks(FROZEN_TIME, dt.timedelta(days=7))
     assert not await state.task_exists(ULID1)
     assert await state.task_exists(ULID2)
 
 
-# ── stage_requeue / stage_submit_task / stage_remove_from_queue ───────────────
-# These three tests use state.pipeline() (not data_store.pipeline()) so
-# they work against all three backends.
+# ── pipeline staging (AtomicTaskStateProtocol) ────────────────────────────────
+# Uses state.pipeline() so these work against all three backends.
 
 
 @pytest.mark.asyncio
 async def test_stage_requeue_saves_task_data(task_adapter):
-    """stage_requeue also persists the task blob so it can be retrieved."""
+    """stage_requeue persists the task blob so it can be retrieved."""
     state, submit = task_adapter
     task = make_task()
     pipe = state.pipeline(transaction=True)
     state.stage_requeue(pipe, task)
     await pipe.execute()
-
     saved = await state.get_task(ULID1)
     assert saved is not None
     assert saved.id == ULID1
@@ -729,7 +611,6 @@ async def test_stage_submit_task_saves_task_data(task_adapter):
     pipe = state.pipeline(transaction=True)
     state.stage_submit_task(pipe, task)
     await pipe.execute()
-
     saved = await state.get_task(ULID1)
     assert saved is not None
     assert saved.id == ULID1
@@ -739,28 +620,9 @@ async def test_stage_submit_task_saves_task_data(task_adapter):
 async def test_stage_remove_from_queue_noop_when_absent(task_adapter):
     """stage_remove_from_queue does not raise when the task is not in the queue."""
     state, submit = task_adapter
-    task = make_task()
     pipe = state.pipeline(transaction=True)
-    state.stage_remove_from_queue(pipe, task)
+    state.stage_remove_from_queue(pipe, make_task())
     await pipe.execute()  # should not raise
-
-
-# ── get_next_task ─────────────────────────────────────────────────────────────
-
-
-@pytest.mark.asyncio
-async def test_get_next_task_returns_submitted_task(task_adapter):
-    """get_next_task pops and returns the next available task."""
-    state, submit = task_adapter
-    task = make_task()
-    await submit.submit_task(task)
-
-    result = await submit.get_next_task(queues={"default"}, pop_timeout=1)
-    assert result is not None
-    assert result.id == ULID1
-
-    result2 = await submit.get_next_task(queues={"default"}, pop_timeout=1)
-    assert result2 is None
 
 
 # ── ensure_index ──────────────────────────────────────────────────────────────
@@ -780,20 +642,15 @@ async def test_ensure_index_does_not_raise(task_adapter):
 async def test_get_tasks_bulk_empty_input(task_adapter):
     """get_tasks_bulk returns an empty list when given no IDs."""
     state, submit = task_adapter
-    result = await state.get_tasks_bulk([])
-    assert result == []
+    assert await state.get_tasks_bulk([]) == []
 
 
 @pytest.mark.asyncio
 async def test_get_tasks_bulk_returns_tasks_in_input_order(task_adapter):
     """get_tasks_bulk returns tasks in the same order as the input ID list."""
     state, submit = task_adapter
-    t1 = make_task(ULID1, name="task_a")
-    t2 = make_task(ULID2, name="task_b")
-    t3 = make_task(ULID3, name="task_c")
-    for t in (t1, t2, t3):
+    for t in (make_task(ULID1, name="a"), make_task(ULID2, name="b"), make_task(ULID3, name="c")):
         await state.save_task(t)
-
     result = await state.get_tasks_bulk([ULID3, ULID1, ULID2])
     assert [r.id for r in result] == [ULID3, ULID1, ULID2]
 
@@ -803,7 +660,6 @@ async def test_get_tasks_bulk_returns_none_for_missing_ids(task_adapter):
     """get_tasks_bulk returns None in the result list for IDs with no stored task."""
     state, submit = task_adapter
     await state.save_task(make_task(ULID1))
-
     result = await state.get_tasks_bulk([ULID1, ULID2])
     assert result[0] is not None
     assert result[0].id == ULID1
@@ -818,7 +674,6 @@ async def test_get_tasks_bulk_populates_heartbeat_at(task_adapter):
     await state.save_task(task)
     task.heartbeat_at = FROZEN_TIME
     await state.update_task_heartbeat(task)
-
     result = await state.get_tasks_bulk([ULID1])
     assert len(result) == 1
     assert result[0] is not None
@@ -841,16 +696,13 @@ async def test_get_dag_runs_returns_empty_when_none(task_adapter):
 async def test_get_dag_runs_returns_submitted_dag_runs(task_adapter):
     """get_dag_runs returns registered DAG run IDs with their submission timestamps."""
     state, submit = task_adapter
-    dag_run_id = ULID1
     task = make_task()
-    task.dag_run_id = dag_run_id
+    task.dag_run_id = ULID1
     await submit.submit_task(task)
-
     runs, total = await state.get_dag_runs(DAGRunPagination())
     assert total == 1
-    assert len(runs) == 1
     run_id, submitted_at = runs[0]
-    assert run_id == dag_run_id
+    assert run_id == ULID1
     assert submitted_at == pytest.approx(FROZEN_TIME, abs=dt.timedelta(seconds=1))
 
 
@@ -858,8 +710,7 @@ async def test_get_dag_runs_returns_submitted_dag_runs(task_adapter):
 async def test_get_dag_run_returns_none_for_unknown(task_adapter):
     """get_dag_run returns None when the dag_run_id has never been registered."""
     state, submit = task_adapter
-    result = await state.get_dag_run(ULID())
-    assert result is None
+    assert await state.get_dag_run(ULID()) is None
 
 
 @pytest.mark.asyncio
@@ -873,9 +724,7 @@ async def test_get_dag_run_returns_submitted_at_and_task_ids(task_adapter):
     task_b.dag_run_id = dag_run_id
     await submit.submit_task(task_a)
     await submit.submit_task(task_b)
-
     result = await state.get_dag_run(dag_run_id)
-
     assert result is not None
     submitted_at, task_ids = result
     assert submitted_at == pytest.approx(FROZEN_TIME, abs=dt.timedelta(seconds=1))
@@ -889,13 +738,10 @@ async def test_get_dag_run_returns_submitted_at_and_task_ids(task_adapter):
 async def test_clean_dag_runs_removes_old_entries(task_adapter):
     """clean_dag_runs removes DAG run entries older than max_age."""
     state, submit = task_adapter
-    dag_run_id = ULID1
     old_task = make_task(submitted_at=FROZEN_TIME - dt.timedelta(days=10))
-    old_task.dag_run_id = dag_run_id
+    old_task.dag_run_id = ULID1
     await submit.submit_task(old_task)
-
     await state.clean_dag_runs(FROZEN_TIME, dt.timedelta(days=7))
-
     runs, total = await state.get_dag_runs(DAGRunPagination())
     assert total == 0
     assert runs == []
@@ -905,16 +751,13 @@ async def test_clean_dag_runs_removes_old_entries(task_adapter):
 async def test_clean_dag_runs_keeps_recent_entries(task_adapter):
     """clean_dag_runs leaves recent DAG run entries untouched."""
     state, submit = task_adapter
-    dag_run_id = ULID1
     recent_task = make_task(submitted_at=FROZEN_TIME - dt.timedelta(hours=1))
-    recent_task.dag_run_id = dag_run_id
+    recent_task.dag_run_id = ULID1
     await submit.submit_task(recent_task)
-
     await state.clean_dag_runs(FROZEN_TIME, dt.timedelta(days=7))
-
     runs, total = await state.get_dag_runs(DAGRunPagination())
     assert total == 1
-    assert runs[0][0] == dag_run_id
+    assert runs[0][0] == ULID1
 
 
 # ── fan-in ────────────────────────────────────────────────────────────────────
@@ -927,9 +770,7 @@ async def test_init_fan_in_creates_expected_members(task_adapter):
     fan_in_key = "fan-in:init-test"
     predecessor_ids = {ULID1, ULID2, ULID3}
     await state.init_fan_in(fan_in_key, predecessor_ids)
-
-    result = await state.get_fan_in_members(fan_in_key)
-    assert set(result) == predecessor_ids
+    assert set(await state.get_fan_in_members(fan_in_key)) == predecessor_ids
 
 
 @pytest.mark.asyncio
@@ -938,23 +779,16 @@ async def test_fan_in_complete_returns_remaining_count(task_adapter):
     state, submit = task_adapter
     fan_in_key = "fan-in:countdown"
     await state.init_fan_in(fan_in_key, {ULID1, ULID2})
-
-    remaining = await state.fan_in_complete(fan_in_key, ULID1)
-    assert remaining == 1
-
-    remaining = await state.fan_in_complete(fan_in_key, ULID2)
-    assert remaining == 0
+    assert await state.fan_in_complete(fan_in_key, ULID1) == 1
+    assert await state.fan_in_complete(fan_in_key, ULID2) == 0
 
 
 @pytest.mark.asyncio
 async def test_fan_in_complete_returns_minus_one_for_unknown_id(task_adapter):
     """fan_in_complete returns -1 when the task ID is not a member of the set."""
     state, submit = task_adapter
-    fan_in_key = "fan-in:unknown"
-    await state.init_fan_in(fan_in_key, {ULID1})
-
-    result = await state.fan_in_complete(fan_in_key, ULID2)
-    assert result == -1
+    await state.init_fan_in("fan-in:unknown", {ULID1})
+    assert await state.fan_in_complete("fan-in:unknown", ULID2) == -1
 
 
 @pytest.mark.asyncio
@@ -964,10 +798,6 @@ async def test_get_fan_in_members_returns_predecessor_ids(task_adapter):
     fan_in_key = "fan-in:members-test"
     predecessor_ids = {ULID1, ULID2, ULID3}
     await state.init_fan_in(fan_in_key, predecessor_ids)
-
-    # exhaust the tracking set so it's empty, but members should still be retrievable
     for uid in predecessor_ids:
         await state.fan_in_complete(fan_in_key, uid)
-
-    result = await state.get_fan_in_members(fan_in_key)
-    assert set(result) == predecessor_ids
+    assert set(await state.get_fan_in_members(fan_in_key)) == predecessor_ids
